// src/app/services/openapi-client/model/branchDto.ts
/**
 * ShopErpApi | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BranchDto { 
    id?: number;
    name?: string;
    code?: string;
    address?: string | null;
    phone?: string | null;
    email?: string | null;
    isMain?: boolean | null;
    isActive?: boolean | null;
    createdAt?: string | null;
    updatedAt?: string | null;
}