// src/app/services/openapi-client/model/updateBranchDto.ts
/**
 * ShopErpApi | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateBranchDto { 
    name: string;
    address?: string | null;
    phone?: string | null;
    email?: string | null;
    isActive?: boolean;
}