// src/app/services/openapi-client/model/branchDtoIEnumerableApiResponseDto.ts
/**
 * ShopErpApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BranchDto } from './branchDto';


export interface BranchDtoIEnumerableApiResponseDto { 
    success?: boolean;
    message?: string | null;
    data?: Array<BranchDto> | null;
    errors?: Array<string> | null;
}