// src/app/services/openapi-client/model/branchDtoApiResponseDto.ts
/**
 * ShopErpApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BranchDto } from './branchDto';


export interface BranchDtoApiResponseDto { 
    success?: boolean;
    message?: string | null;
    data?: BranchDto;
    errors?: Array<string> | null;
}