// src/app/services/openapi-client/model/booleanApiResponseDto.ts
/**
 * ShopErpApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BooleanApiResponseDto { 
    success?: boolean;
    message?: string | null;
    data?: boolean;
    errors?: Array<string> | null;
}